Add pkgconf to deps
For some reason choco install is very slow. And explicitly specifying the
version is good for reproducibility anyway.

There are no obviously trustworth places (it's typically old sourceforge
pages) to download pkgconf. So it seems easiest to just build it ourselves.

If desirable I can add a matrix to postgresql-dev showing what can be built
with and without pkgconf.